backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend template with FastAPI + SQLAlchemy + Alembic + PostgreSQL + Redis + Celery"
authors = []
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { version = "^0.34.0", extras = ["standard"] }
sqlalchemy = { version = "^2.0.0", extras = ["asyncio"] }
asyncpg = "^0.30.0"
alembic = "^1.14.0"
redis = { version = "^5.0.0", extras = ["hiredis"] }
celery = { version = "^5.4.0", extras = ["redis"] }
pydantic-settings = "^2.0.0"
passlib = { version = "^1.7.4", extras = ["bcrypt"] }
bcrypt = "^4.0.0"
email-validator = "^2.0.0"
openai = "^2.0.0"
httpx = "^0.28.0"
python-multipart = "^0.0.18"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
